Refactor explorer API watchdogs
Use the same actor for all explorers we'll want to support.
Each explorer watchdog simply implements the API calls.

Fix the `BlockchainWatchdogSpec`: there was a race condition where we may
send `CurrentBlockCount` to the `eventStream` before the actor under test was
registered to these events.
